Tests/LibWeb/Text/expected/css/omitted-circle-function-position-interpolation.txt

circle(15px)
circle(at 100% 100%)
circle()
circle(15px at 25% 25%)
